refactor(web): app router 방식으로 페이지 구조 변경
