+++ apps/web/src/_pages/landing/index.ts
@@ -1,0 +1,4 @@
+export * from "./assets";
+export { CATEGORIES } from "./constants";
+export { CategoryCard } from "./ui/category-card";
+export { IcoChevronDownDouble } from "./ui/ico-chevron-down-double";
